explorerd

Module error

Crate errors

Enums

ExplorerdError

Custom RPC errors available for blockchain explorer. These represent specific RPC-related failures.

Constants

ERROR_CODE_PING_DARKFID_FAILED

Functions

handle_database_error

Logs and converts a database error into a [DatabaseError]. This function ensures the error is logged contextually before being returned.

server_error

Constructs a JsonError representing a server error using the provided ExplorerdError , request ID, and optional custom message, returning a JsonError with a corresponding server error code and message.

to_error_code_message

Helper function to convert ExplorerdRpcError into error code with corresponding error message.