explorerd::service

Module contracts

Implements functionality for smart contracts

Structs

ContractRecord

Represents a contract record embellished with details that are not stored on-chain.

Functions

untar_source

Auxiliary function that extracts source code files from a TAR archive provided as a byte slice [&[u8]], returning a Vec of ContractSourceFiles representing the extracted file paths and their contents.