explorerd::store
Module contract_metadata

Stores, manages, and provides access to contract metadata

Structs

ContractMetaData

Represents contract metadata containing additional contract information that is not stored on-chain.

ContractMetaStore

ContractMetadataStoreOverlay 🔒

The ContractMetadataStoreOverlay provides write operations for managing contract metadata in underlying sled database. It supports inserting new ContractMetaData and contract source code ContractSourceFile content and deleting existing source code.

ContractSourceFile

Represents a source file containing its file path as a string and its content as a vector of bytes.

Constants

SLED_CONTRACT_METADATA_TREE

Contract metadata tree name.

SLED_CONTRACT_SOURCE_CODE_TREE

Contract source code tree name.